1. base aio_tcp_transport implement http_session 2. http_session use reference count

+++ libhttp/include/http-bundle.h
@@ -3,7 +3,7 @@
 
 struct http_bundle_t
 {
-	long ref;
+	int32_t ref;
 
 	size_t len;  // used size
 	void* ptr;
